[REFACTOR] Rename stages to user
