+++ src/client/react/user/Promotion/index.js
