+++ src/client/react/user/Registration/Dashboard/index.js
