+++ src/client/react/user/Registration/Home.js
