+++ src/client/react/user/Registration/RegisterForm/Form2.js
