+++ src/client/react/user/Registration/RegisterForm/index.js
