+++ src/client/react/user/Registration/index.js
