+++ src/client/react/user/RegistrationClosed/Dashboard.js
