+++ src/client/react/user/RegistrationClosed/Home.js
