+++ src/client/react/user/RegistrationClosed/index.js
