+++ src/client/react/user/components/Pay.js
