+++ src/client/react/user/pages/Home/TeamPanel.js
