+++ src/client/react/user/pages/Instructions/InstructionCollapse.js
