+++ src/client/react/user/pages/Instructions/index.js
